+++ detector/src/main/java/com/back/detector/domain/CrawlingDetectResult.java
@@ -2,6 +2,6 @@ package com.back.detector.domain;
 
 public record CrawlingDetectResult(
         CrawlingBanLevel banLevel,
-        int requestCount
+        long requestCount
 ) {
 }
