+++ detector/src/main/java/com/back/detector/mapper/CrawlingLogMapper.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class CrawlingLogMapper {
 
-    public CrawlingLog toCrawlingLog(String ip, int requestCount, CrawlingBanLevel banLevel) {
+    public CrawlingLog toCrawlingLog(String ip, long requestCount, CrawlingBanLevel banLevel) {
         return CrawlingLog.builder()
                 .ipAddress(ip)
                 .requestCount(requestCount)
